# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout user",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Logout successful"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Get current user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user data"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "role"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid input or user already exists"
          }
        }
      }
    },
    "/api/cases/{id}": {
      "get": {
        "summary": "Get case by ID",
        "tags": [
          "Cases"
        ],
        "responses": {
          "200": {
            "description": "Case details"
          }
        }
      },
      "put": {
        "summary": "Update case",
        "tags": [
          "Cases"
        ],
        "responses": {
          "200": {
            "description": "Case updated"
          }
        }
      },
      "delete": {
        "summary": "Delete case",
        "tags": [
          "Cases"
        ],
        "responses": {
          "200": {
            "description": "Case deleted"
          }
        }
      }
    },
    "/api/cases/attorney/{userId}": {
      "get": {
        "summary": "Get cases assigned to attorney",
        "tags": [
          "Cases"
        ],
        "responses": {
          "200": {
            "description": "List of attorney cases"
          }
        }
      }
    },
    "/api/cases": {
      "get": {
        "summary": "Get all cases",
        "tags": [
          "Cases"
        ],
        "responses": {
          "200": {
            "description": "List of cases"
          }
        }
      },
      "post": {
        "summary": "Create a new case",
        "tags": [
          "Cases"
        ],
        "responses": {
          "201": {
            "description": "Case created"
          }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "summary": "Get customer by ID",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details"
          }
        }
      },
      "put": {
        "summary": "Update customer",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer updated"
          }
        }
      },
      "delete": {
        "summary": "Delete customer",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted"
          }
        }
      }
    },
    "/api/customers": {
      "get": {
        "summary": "Get all customers",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created"
          }
        }
      }
    },
    "/api/example": {
      "get": {
        "summary": "Example GET endpoint",
        "description": "Returns a sample response to demonstrate API structure",
        "tags": [
          "Examples"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Hello from the API"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "success"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Example POST endpoint",
        "description": "Accepts data and returns a confirmation",
        "tags": [
          "Examples"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data received successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "receivedData": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    }
  },
  "tags": []
}